fix: error handling (all pipelines + google)
Now that we can edit custom pipelines options, it really makes sense to
put the timedObservationsViewer as customPipelineOptions instead of
having it at the root of the manifest. Thanks to this, custom pipelines
no longer need to pass the entire manifest.

Also improves error handling in general for all pipelines.